refactor: Moving client and configuration to be more inline with other plugins e.g. Gandi
This mostly was about breaking out the existing `Client` struct into a googleanalytics service and a plugin client, which is typically how it is done in other plugins.
