+++ plugins/source/googleanalytics/resources/plugin/plugin.go
@@ -1,7 +1,6 @@
 package plugin
 
 import (
-	"github.com/cloudquery/cloudquery/plugins/source/googleanalytics/client"
 	"github.com/cloudquery/plugin-sdk/v4/plugin"
 )
 
@@ -16,7 +15,7 @@ func Plugin() *plugin.Plugin {
 	return plugin.NewPlugin(
 		Name,
 		Version,
-		client.Configure,
+		Configure,
 		plugin.WithKind(Kind),
 		plugin.WithTeam(Team),
 	)
